Fix examples

+++ cpu-dual/cpu-dual.c
@@ -33,10 +33,10 @@ main(void)
         assert((cpu_dual_executor_get()) == CPU_MASTER);
 
         dbgio_printf("Master stack address: 0x%08lX\n",
-            (uint32_t)cpu_dual_master_stack_get());
+            (uintptr_t)cpu_dual_master_stack_get());
 
         dbgio_printf("Slave stack address: 0x%08lX\n",
-            (uint32_t)cpu_dual_slave_stack_get());
+            (uintptr_t)cpu_dual_slave_stack_get());
 
         dbgio_puts("Ping ponging between master and slave\n");
